requirements.txt
fastapi
uvicorn
selenium
beautifulsoup4
pymongo
python-dotenv
webdriver_manager